requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.27
langchain-community==0.3.31
chromadb==1.1.1
sentence-transformers==5.1.1
pymupdf==1.26.5
tiktoken==0.12.0
huggingface_hub==0.35.3
llama-cpp-python==0.2.28
numpy==2.3.3
pydantic>=2.0
